Bug 1307822 - Calls the sync completion block on background to be on the main thread
A potential fix for our springboard/disk access crash since in all of
the crash logs we seem to be running the sqlite3_close->WAL
checkpointing on a background thread which could be racing against the
app closing. Another potential cause could be that the background
identifier we get back from UIApplication is not thread-aware.